requirements.txt
email-validator==2.0.0.post2
fastapi==0.96.0
pydantic==1.10.9
SQLAlchemy==2.0.15
typing-extensions==4.6.3
uvicorn==0.22.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6


